Abstracting room utlity functions from roomController.ts
